pyproject.toml
[project]
name = "illumio-mcp"
version = "0.1.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "illumio>=1.1.3",
 "logging>=0.4.9.6",
 "mcp>=1.2.0",
 "pandas>=2.2.3",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "alexgoller"
email = "83632450+alexgoller@users.noreply.github.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
illumio-mcp = "illumio_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/illumio_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
